+++ src/main/java/com/ing/data/cassandra/jdbc/CassandraConnection.java
@@ -130,7 +130,7 @@ public class CassandraConnection extends AbstractConnection implements Connectio
     @SuppressWarnings("SortedCollectionWithNonComparableKeys")
     private final Set<Statement> statements = new ConcurrentSkipListSet<>();
     private final ConcurrentMap<String, CassandraPreparedStatement> preparedStatements = new ConcurrentHashMap<>();
-    private final ConsistencyLevel defaultConsistencyLevel;
+    private ConsistencyLevel consistencyLevel;
     private int defaultFetchSize = FALLBACK_FETCH_SIZE;
     private String currentKeyspace;
     private final boolean debugMode;
@@ -157,7 +157,7 @@ public class CassandraConnection extends AbstractConnection implements Connectio
         this.optionSet = lookupOptionSet(sessionProperties.getProperty(TAG_COMPLIANCE_MODE));
         this.username = sessionProperties.getProperty(TAG_USER,
             defaultConfigProfile.getString(DefaultDriverOption.AUTH_PROVIDER_USER_NAME, StringUtils.EMPTY));
-        this.defaultConsistencyLevel = DefaultConsistencyLevel.valueOf(
+        this.consistencyLevel = DefaultConsistencyLevel.valueOf(
             sessionProperties.getProperty(TAG_CONSISTENCY_LEVEL,
                 defaultConfigProfile.getString(DefaultDriverOption.REQUEST_CONSISTENCY,
                     ConsistencyLevel.LOCAL_ONE.name())));
@@ -219,7 +219,7 @@ public class CassandraConnection extends AbstractConnection implements Connectio
         this.currentKeyspace = currentKeyspace;
         this.cSession = cSession;
         this.metadata = cSession.getMetadata();
-        this.defaultConsistencyLevel = defaultConsistencyLevel;
+        this.consistencyLevel = defaultConsistencyLevel;
         this.debugMode = debugMode;
         final List<TypeCodec<?>> codecs = new ArrayList<>();
         codecs.add(new TimestampToLongCodec());
@@ -375,12 +375,21 @@ public class CassandraConnection extends AbstractConnection implements Connectio
     }
 
     /**
-     * Gets the default consistency level applied to this connection.
+     * Gets the consistency level currently applied to this connection.
      *
-     * @return The default consistency level applied to this connection.
+     * @return The consistency level currently applied to this connection.
      */
-    public ConsistencyLevel getDefaultConsistencyLevel() {
-        return this.defaultConsistencyLevel;
+    public ConsistencyLevel getConsistencyLevel() {
+        return this.consistencyLevel;
+    }
+
+    /**
+     * Sets the consistency level applied to this connection.
+     *
+     * @param consistencyLevel  The consistency level to apply to this connection.
+     */
+    public void setConsistencyLevel(final ConsistencyLevel consistencyLevel) {
+         this.consistencyLevel = consistencyLevel;
     }
 
     /**