+++ src/main/java/com/ing/data/cassandra/jdbc/codec/BigintToBigDecimalCodec.java
@@ -22,7 +22,7 @@ import com.datastax.oss.driver.api.core.type.codec.TypeCodec;
 import com.datastax.oss.driver.api.core.type.reflect.GenericType;
 import com.ing.data.cassandra.jdbc.utils.ByteBufferUtil;
 
-import javax.annotation.Nonnull;
+import jakarta.annotation.Nonnull;
 import java.math.BigDecimal;
 import java.nio.ByteBuffer;